refactor: Creates separate component for inputs and buttons on chat
