+++ src/components/ChatAction/index.js
@@ -1,0 +1,3 @@
+import ChatAction from "./ChatAction";
+
+export default ChatAction
