+++ src/components/index.js
@@ -1,5 +1,7 @@
 import ChatBox from "./ChatBox";
+import ChatAction from "./ChatAction";
 
 export {
-  ChatBox
+  ChatBox,
+  ChatAction,
 }
